+++ source/Makefile.Objects
@@ -725,6 +725,7 @@ OBJS_SRCPW=h_ewald_pw.o\
     vnl_pw.o\
     vnl_pw_getvnl.o\
     vnl_pw_grad.o\
+    vnl_pw_qrad.o\
     chgmixing.o\
     charge.o\
     charge_init.o\
