+++ source/source_pw/module_pwdft/CMakeLists.txt
@@ -45,6 +45,7 @@ list(APPEND objects
     vnl_pw.cpp
     vnl_pw_getvnl.cpp
     vnl_pw_grad.cpp
+    vnl_pw_qrad.cpp
     stru_fac.cpp
     stru_fac_k.cpp
     soc.cpp
