+++ source/source_pw/module_pwdft/vnl_pw.cpp
@@ -674,238 +674,6 @@ void pseudopot_cell_vnl::init_vnl(UnitCell& cell, const ModulePW::PW_Basis* rho_
     return;
 }
 
-void pseudopot_cell_vnl::compute_qrad(UnitCell& cell)
-{
-    const double pref = ModuleBase::FOUR_PI / cell.omega;
-
-    for (int it = 0; it < cell.ntype; it++)
-    {
-        Atom_pseudo* upf = &cell.atoms[it].ncpp;
-        if (upf->tvanp)
-        {
-            const int nbeta = upf->nbeta;
-            int kkbeta = upf->kkbeta;
-            if ((kkbeta % 2 == 0) && kkbeta > 0)
-            {
-                kkbeta--;
-            }
-            double* aux = new double[kkbeta];
-            double* besr = new double[kkbeta];
-
-            for (int l = 0; l < upf->nqlc; l++)
-            {
-                for (int iq = 0; iq < PARAM.globalv.nqxq; iq++)
-                {
-                    const double q = iq * PARAM.globalv.dq;
-                    // here we compute the spherical bessel function for each q_i
-                    ModuleBase::Sphbes::Spherical_Bessel(kkbeta, upf->r.data(), q, l, besr);
-                    for (int nb = 0; nb < nbeta; nb++)
-                    {
-                        // the Q are symmetric with respect to indices nb and mb
-                        for (int mb = nb; mb < nbeta; mb++)
-                        {
-                            const int ijv = mb * (mb + 1) / 2 + nb;
-                            if ((l >= std::abs(upf->lll[nb] - upf->lll[mb])) && (l <= (upf->lll[nb] + upf->lll[mb]))
-                                && ((l + upf->lll[nb] + upf->lll[mb]) % 2 == 0))
-                            {
-                                for (int ir = 0; ir < kkbeta; ir++)
-                                {
-                                    aux[ir] = besr[ir] * upf->qfuncl(l, ijv, ir);
-                                }
-                                // then we integrate with all the Q functions
-                                double vqint = 0.0;
-                                ModuleBase::Integral::Simpson_Integral(kkbeta, aux, upf->rab.data(), vqint);
-                                qrad(it, l, ijv, iq) = vqint * pref;
-                            }
-                        }
-                    }
-                }
-            }
-            delete[] aux;
-            delete[] besr;
-        }
-    }
-}
-
-void pseudopot_cell_vnl::radial_fft_q(const int ng,
-                                      const int ih,
-                                      const int jh,
-                                      const int itype,
-                                      const double* qnorm,
-                                      const ModuleBase::matrix ylm,
-                                      std::complex<double>* qg) const
-{
-    // computes the indices which correspond to ih,jh
-    const int nb = indv(itype, ih);
-    const int mb = indv(itype, jh);
-    assert(nb < nbetam);
-    assert(mb < nbetam);
-    int ijv = 0;
-    if (nb >= mb)
-    {
-        ijv = nb * (nb + 1) / 2 + mb;
-    }
-    else
-    {
-        ijv = mb * (mb + 1) / 2 + nb;
-    }
-    const int ivl = nhtolm(itype, ih);
-    const int jvl = nhtolm(itype, jh);
-
-    for (int ig = 0; ig < ng; ig++)
-    {
-        qg[ig] = {0, 0};
-    }
-    // makes the sum over the non zero LM
-    int l = -1;
-    std::complex<double> pref(0.0, 0.0);
-    for (int lm = 0; lm < this->lpx(ivl, jvl); lm++)
-    {
-        int lp = this->lpl(ivl, jvl, lm);
-        assert(lp >= 0);
-        assert(lp < 49);
-        if (lp == 0)
-        {
-            l = 0;
-        }
-        else if (lp < 4)
-        {
-            l = 1;
-        }
-        else if (lp < 9)
-        {
-            l = 2;
-        }
-        else if (lp < 16)
-        {
-            l = 3;
-        }
-        else if (lp < 25)
-        {
-            l = 4;
-        }
-        else if (lp < 36)
-        {
-            l = 5;
-        }
-        else
-        {
-            l = 6;
-        }
-        pref = pow(ModuleBase::NEG_IMAG_UNIT, l) * this->ap(lp, ivl, jvl);
-
-        double qm1 = -1.0; // any number smaller than qnorm
-        double work = 0.0;
-        for (int ig = 0; ig < ng; ig++)
-        {
-            // calculate quantites depending on the module of G only when needed
-            if (std::abs(qnorm[ig] - qm1) > 1e-6)
-            {
-                work = ModuleBase::PolyInt::Polynomial_Interpolation(this->qrad,
-                                                                     itype,
-                                                                     l,
-                                                                     ijv,
-                                                                     PARAM.globalv.nqxq,
-                                                                     PARAM.globalv.dq,
-                                                                     qnorm[ig]);
-                qm1 = qnorm[ig];
-            }
-            qg[ig] += pref * work * ylm(lp, ig);
-        }
-    }
-}
-
-template <typename FPTYPE, typename Device>
-void pseudopot_cell_vnl::radial_fft_q(Device* ctx,
-                                      const int ng,
-                                      const int ih,
-                                      const int jh,
-                                      const int itype,
-                                      const FPTYPE* qnorm,
-                                      const FPTYPE* ylm,
-                                      std::complex<FPTYPE>* qg) const
-{
-    using setmem_complex_op = base_device::memory::set_memory_op<std::complex<FPTYPE>, Device>;
-
-    // computes the indices which correspond to ih,jh
-    const int nb = indv(itype, ih);
-    const int mb = indv(itype, jh);
-    assert(nb < nbetam);
-    assert(mb < nbetam);
-    int ijv = 0;
-    if (nb >= mb)
-    {
-        ijv = nb * (nb + 1) / 2 + mb;
-    }
-    else
-    {
-        ijv = mb * (mb + 1) / 2 + nb;
-    }
-    const int ivl = nhtolm(itype, ih);
-    const int jvl = nhtolm(itype, jh);
-
-    setmem_complex_op()(qg, 0, ng);
-
-    const double* qnorm_double = reinterpret_cast<const double*>(qnorm);
-
-    // makes the sum over the non zero LM
-    int l = -1;
-    std::complex<FPTYPE> pref(0.0, 0.0);
-    for (int lm = 0; lm < this->lpx(ivl, jvl); lm++)
-    {
-        int lp = this->lpl(ivl, jvl, lm);
-        assert(lp >= 0);
-        assert(lp < 49);
-        if (lp == 0)
-        {
-            l = 0;
-        }
-        else if (lp < 4)
-        {
-            l = 1;
-        }
-        else if (lp < 9)
-        {
-            l = 2;
-        }
-        else if (lp < 16)
-        {
-            l = 3;
-        }
-        else if (lp < 25)
-        {
-            l = 4;
-        }
-        else if (lp < 36)
-        {
-            l = 5;
-        }
-        else
-        {
-            l = 6;
-        }
-        pref = static_cast<std::complex<FPTYPE>>(pow(ModuleBase::NEG_IMAG_UNIT, l) * this->ap(lp, ivl, jvl));
-
-        double qm1 = -1.0; // any number smaller than qnorm
-        double work = 0.0;
-        for (int ig = 0; ig < ng; ig++)
-        {
-            if (std::abs(qnorm_double[ig] - qm1) > 1e-6)
-            {
-                work = ModuleBase::PolyInt::Polynomial_Interpolation(this->qrad,
-                                                                     itype,
-                                                                     l,
-                                                                     ijv,
-                                                                     PARAM.globalv.nqxq,
-                                                                     PARAM.globalv.dq,
-                                                                     qnorm_double[ig]);
-                qm1 = qnorm_double[ig];
-            }
-            qg[ig] += pref * static_cast<FPTYPE>(work) * ylm[lp * ng + ig];
-        }
-    }
-}
-
 #ifdef __LCAO
 std::complex<double> pseudopot_cell_vnl::Cal_C(int alpha, int lu, int mu, int L, int M) // pengfei Li  2018-3-23
 {
@@ -1604,38 +1372,3 @@ std::complex<double>* pseudopot_cell_vnl::get_qq_so_data() const
 {
     return this->z_qq_so;
 }
-
-template void pseudopot_cell_vnl::radial_fft_q<float, base_device::DEVICE_CPU>(base_device::DEVICE_CPU*,
-                                                                               const int,
-                                                                               const int,
-                                                                               const int,
-                                                                               const int,
-                                                                               const float*,
-                                                                               const float*,
-                                                                               std::complex<float>*) const;
-template void pseudopot_cell_vnl::radial_fft_q<double, base_device::DEVICE_CPU>(base_device::DEVICE_CPU*,
-                                                                                const int,
-                                                                                const int,
-                                                                                const int,
-                                                                                const int,
-                                                                                const double*,
-                                                                                const double*,
-                                                                                std::complex<double>*) const;
-#if defined(__CUDA) || defined(__ROCM)
-template void pseudopot_cell_vnl::radial_fft_q<float, base_device::DEVICE_GPU>(base_device::DEVICE_GPU*,
-                                                                               const int,
-                                                                               const int,
-                                                                               const int,
-                                                                               const int,
-                                                                               const float*,
-                                                                               const float*,
-                                                                               std::complex<float>*) const;
-template void pseudopot_cell_vnl::radial_fft_q<double, base_device::DEVICE_GPU>(base_device::DEVICE_GPU*,
-                                                                                const int,
-                                                                                const int,
-                                                                                const int,
-                                                                                const int,
-                                                                                const double*,
-                                                                                const double*,
-                                                                                std::complex<double>*) const;
-#endif
